// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  description?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Footer {
  id: number;
  logo: number | Media;
  nav: {
    label?: string | null;
    link?: string | null;
    id?: string | null;
  }[];
  copyright: string;
  updatedAt?: string | null;
  createdAt?: string | null;
}